requirements.txt
opencv-python~=4.5.0
scikit-image
tensorflow~=2.8.0
pillow
matplotlib